extract idle hardware
