+++ firmware/controllers/actuators/idle_thread.h
@@ -35,10 +35,6 @@ void setIdleDFactor(float value);
 void setIdleMode(idle_mode_e value DECLARE_ENGINE_PARAMETER_SUFFIX);
 void setTargetIdleRpm(int value);
 void setIdleDT(int value);
-void stopIdleHardware(DECLARE_ENGINE_PARAMETER_SIGNATURE);
-void initIdleHardware(DECLARE_ENGINE_PARAMETER_SIGNATURE);
-bool isIdleHardwareRestartNeeded();
-bool isIdleMotorBusy(DECLARE_ENGINE_PARAMETER_SIGNATURE);
 void onConfigurationChangeIdleCallback(engine_configuration_s *previousConfiguration);
 float getIdlePidOffset(DECLARE_ENGINE_PARAMETER_SIGNATURE);
 Pid * getIdlePid(DECLARE_ENGINE_PARAMETER_SIGNATURE);
